Moves middleware into photo and text dirs
