+++ lib/middleware/campaignActivity/photo/draft-caption.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const helpers = require('../../helpers');
-const reportbackHelper = require('../../helpers/reportback');
-const replies = require('../../replies');
+const helpers = require('../../../helpers');
+const reportbackHelper = require('../../../helpers/reportback');
+const replies = require('../../../replies');
 
 module.exports = function draftSubmissionCaption() {
   return (req, res, next) => {
