+++ lib/middleware/campaignActivity/photo/draft-why-participated.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const helpers = require('../../helpers');
-const reportbackHelper = require('../../helpers/reportback');
-const replies = require('../../replies');
+const helpers = require('../../../helpers');
+const reportbackHelper = require('../../../helpers/reportback');
+const replies = require('../../../replies');
 
 module.exports = function draftSubmissionWhyParticipated() {
   return (req, res, next) => {
