+++ lib/middleware/campaignActivity/text/post-create.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const logger = require('winston');
-const helpers = require('../../helpers');
-const replies = require('../../replies');
+const helpers = require('../../../helpers');
+const replies = require('../../../replies');
 
 module.exports = function textPost() {
   return (req, res, next) => {
